Skeleton support for dynamic files
They now show up in the UI as JS files. At some point there should be a visible
indcation of what files are dynamic and which are standard javascript.

Now, when a dynamic file is reqested, it is evaluated first.

The ajax code will hook into this new implmementation of dynamic files
integrated with the new Backbone classes.