Database in local storage
db.store()

Takes:
A string with the storage name for the DB

Returns:
true if data was returned from localStorage, false if no data returned
Sets up a localStorage collection for a given name. If data exists
already the data will be loaded into the collection. Changes are auto
synced to localStorage.

Pass in false to terminate storing the collection (data in localstorage
will be unchanged).

Note: localStorage is not avaliable in all browsers. If localStorage is
not found then .store() will be ignored.

db.store("name"); // starts storing records in local storage